MAINT: dedicated file for LegacyPlugin
Also: removed ability to register plugins. Instead, directly
register them in imageio.config.
